Print deprecation message only on usage of the returned value
