+++ lib/podlet.js
@@ -4,6 +4,7 @@
 
 'use strict';
 
+const { CssDeprecation, JsDeprecation } = require('./deprecations');
 const { HttpIncoming, pathnameBuilder } = require('@podium/utils');
 const { validate } = require('@podium/schemas');
 const Metrics = require('@metrics/client');
@@ -19,26 +20,6 @@ const _compabillity = Symbol('_compabillity');
 const _sanitize = Symbol('_sanitize');
 const _addCssAsset = Symbol('_addCssAsset');
 const _addJsAsset = Symbol('_addJsAsset');
-
-function deprecateJsReturn() {
-    if (!deprecateJsReturn.warned) {
-        deprecateJsReturn.warned = true;
-        process.emitWarning(
-            'Return value from method js() is now deprecated and will be removed in a future version. Please do not rely on this value.',
-            'DeprecationWarning',
-        );
-    }
-}
-
-function deprecateCssReturn() {
-    if (!deprecateCssReturn.warned) {
-        deprecateCssReturn.warned = true;
-        process.emitWarning(
-            'Return value from method css() is now deprecated and will be removed in a future version. Please do not rely on this value.',
-            'DeprecationWarning',
-        );
-    }
-}
 
 const PodiumPodlet = class PodiumPodlet {
     constructor({
@@ -232,8 +213,7 @@ const PodiumPodlet = class PodiumPodlet {
         });
 
         // deprecate
-        deprecateCssReturn();
-        return this[_sanitize](value, prefix);
+        return new CssDeprecation(this[_sanitize](value, prefix));
     }
 
     css(options = {}) {
@@ -270,8 +250,7 @@ const PodiumPodlet = class PodiumPodlet {
         });
 
         // deprecate
-        deprecateJsReturn();
-        return this[_sanitize](value, prefix);
+        return new JsDeprecation(this[_sanitize](value, prefix));
     }
 
     js(options = {}) {
